# gm-center/gmc_api_gateway/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Gedge GM-Center Swagger API.",
        "title": "Gedge GM-Center Swagger API",
        "contact": {
            "name": "GM-Center",
            "url": "https://gedge-platform.github.io/gm-center/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "101.79.1.138:8012",
    "basePath": "/gmcapi/v2",
    "paths": {
        "/auth": {
            "post": {
                "description": "get JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User Info Body",
                        "name": "authBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/ceph/monit": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get Ceph Dashboard info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ceph"
                ],
                "summary": "Show Ceph Dashboard info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CEPH"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CEPH": {
            "type": "object",
            "properties": {
                "capacity": {
                    "$ref": "#/definitions/model.Capacity"
                },
                "clientPerformance": {
                    "$ref": "#/definitions/model.ClientPerformance"
                },
                "clusterStatus": {
                    "type": "number"
                },
                "hostNum": {
                    "type": "integer"
                },
                "monitor": {
                    "$ref": "#/definitions/model.Monitor"
                },
                "object": {
                    "$ref": "#/definitions/model.Object"
                },
                "osd": {
                    "$ref": "#/definitions/model.OSD"
                },
                "pgStatus": {},
                "pgs_per_osd": {
                    "type": "number"
                },
                "poolNum": {
                    "type": "integer"
                }
            }
        },
...
        "model.Capacity": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                },
                "used": {
                    "type": "number"
                }
            }
        },
        "model.ClientPerformance": {
            "type": "object",
            "properties": {
                "read_bytes_sec": {
                    "type": "number"
                },
                "read_op_per_sec": {
                    "type": "number"
                },
                "write_bytes_sec": {
                    "type": "number"
                },
                "write_op_per_sec": {
                    "type": "number"
                }
            }
        },
...
        "model.Monitor": {
            "type": "object",
            "properties": {
                "quorum": {},
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.OSD": {
            "type": "object",
            "properties": {
                "in": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "up": {
                    "type": "integer"
                }
            }
        },
        "model.Object": {
            "type": "object",
            "properties": {
                "degraded": {
                    "type": "number"
                },
                "healthy": {
                    "type": "number"
                },
                "misplaced": {
                    "type": "number"
                },
                "unfound": {
                    "type": "number"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "string"
                },
                "Password": {
                    "description": "Email    string `json:\"email\"`",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type \\\"Bearer \\\" and then your API Token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}